requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.0
dj-database-url==0.5.0
Django==3.1.7
django-admin==2.0.1
django-allauth==0.44.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-markdown-view==0.0.4
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-six==1.0.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
passlib==1.7.4
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
screen==1.0.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
xlwt==1.3.0
